refactor amp/bento video extensions: * brightcove * jwplayer * video-iframe * video * vimeo * youtube
