+++ extensions/amp-brightcove/1.0/base-element.js
@@ -1,34 +1,16 @@
-import {BentoBrightcove} from './component';
-import {VideoBaseElement} from '../../amp-video/1.0/video-base-element';
-import {createParseAttrsWithPrefix} from '#preact/parse-props';
+import {BaseElement as BentoVideoBaseElement} from '../../amp-video/1.0/base-element';
+import {Component, layoutSizeDefined, props, usesShadowDom} from './element';
 
-export class BaseElement extends VideoBaseElement {}
-
-/** @override */
-BaseElement['Component'] = BentoBrightcove;
+export class BaseElement extends BentoVideoBaseElement {}
 
 /** @override */
-BaseElement['props'] = {
-  'account': {attr: 'data-account'},
-  'autoplay': {attr: 'autoplay', type: 'boolean'},
-  'embed': {attr: 'data-embed', default: 'default'},
-  'player': {
-    attrs: ['data-player', 'data-player-id'],
-    parseAttrs(element) {
-      const {'player': player, 'playerId': playerId} = element.dataset;
-      return player || playerId || 'default';
-    },
-  },
-  'playlistId': {attr: 'data-playlist-id'},
-  'referrer': {attr: 'data-referrer'},
-  'urlParams': createParseAttrsWithPrefix('data-param-'),
-  'videoId': {attr: 'data-video-id'},
-  // TODO(wg-bento): These props have no internal implementation yet.
-  'dock': {attr: 'dock', media: true},
-};
+BaseElement['Component'] = Component;
 
 /** @override */
-BaseElement['layoutSizeDefined'] = true;
+BaseElement['props'] = props;
 
 /** @override */
-BaseElement['usesShadowDom'] = true;
+BaseElement['layoutSizeDefined'] = layoutSizeDefined;
+
+/** @override */
+BaseElement['usesShadowDom'] = usesShadowDom;
