+++ extensions/amp-video-iframe/1.0/base-element.js
@@ -1,31 +1,17 @@
-import {BentoVideoIframe} from './component';
+import {Component, loadable, props, usesShadowDom} from './element';
 
-import {VideoBaseElement} from '../../amp-video/1.0/video-base-element';
+import {BaseElement as BentoVideoBaseElement} from '../../amp-video/1.0/base-element';
 
-export class BaseElement extends VideoBaseElement {}
-
-/** @override */
-BaseElement['Component'] = BentoVideoIframe;
+export class BaseElement extends BentoVideoBaseElement {}
 
 /** @override */
-BaseElement['props'] = {
-  'autoplay': {attr: 'autoplay', type: 'boolean'},
-  'referrerpolicy': {attr: 'referrerpolicy'},
-  'implements-media-session': {attr: 'mediasession', type: 'boolean'},
-  'poster': {attr: 'poster'},
-  'src': {attr: 'src'},
-  'controls': {attr: 'controls', type: 'boolean'},
-  // TODO(wg-components): Current behavior defaults to loading="auto".
-  // Refactor to make loading="lazy" as the default.
-  'loading': {attr: 'data-loading'},
-
-  // TODO(alanorozco): These props have no internal implementation yet.
-  'dock': {attr: 'dock'},
-  'rotate-to-fullscreen': {attr: 'rotate-to-fullscreen', type: 'boolean'},
-};
+BaseElement['Component'] = Component;
 
 /** @override */
-BaseElement['usesShadowDom'] = true;
+BaseElement['props'] = props;
 
 /** @override */
-BaseElement['loadable'] = true;
+BaseElement['usesShadowDom'] = usesShadowDom;
+
+/** @override */
+BaseElement['loadable'] = loadable;
