+++ extensions/amp-video-iframe/1.0/element.js
@@ -1,0 +1,21 @@
+export {BentoVideoIframe as Component} from './component';
+
+export const props = {
+  'autoplay': {attr: 'autoplay', type: 'boolean'},
+  'referrerpolicy': {attr: 'referrerpolicy'},
+  'implements-media-session': {attr: 'mediasession', type: 'boolean'},
+  'poster': {attr: 'poster'},
+  'src': {attr: 'src'},
+  'controls': {attr: 'controls', type: 'boolean'},
+  // TODO(wg-components): Current behavior defaults to loading="auto".
+  // Refactor to make loading="lazy" as the default.
+  'loading': {attr: 'data-loading'},
+
+  // TODO(alanorozco): These props have no internal implementation yet.
+  'dock': {attr: 'dock'},
+  'rotate-to-fullscreen': {attr: 'rotate-to-fullscreen', type: 'boolean'},
+};
+
+export const usesShadowDom = true;
+
+export const loadable = true;
