+++ extensions/amp-video/1.0/amp-video.js
@@ -1,12 +1,12 @@
 import {CSS} from '../../../build/amp-video-1.0.css';
-import {VideoBaseElement} from './video-base-element';
+import {AmpVideoBaseElement} from './video-base-element';
 import {isExperimentOn} from '#experiments';
 import {userAssert} from '#utils/log';
 
 /** @const {string} */
 const TAG = 'amp-video';
 
-class AmpVideo extends VideoBaseElement {
+class AmpVideo extends AmpVideoBaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
