+++ extensions/amp-video/1.0/base-element.js
@@ -1,18 +1,24 @@
-import {CSS} from './component.jss';
-import {CSS as CSS_AUTOPLAY} from './autoplay.jss';
 import {PreactBaseElement} from '#preact/base-element';
-import {BentoVideo} from './component';
+
+import {
+  Component,
+  layoutSizeDefined,
+  loadable,
+  props,
+  shadowCss,
+  usesShadowDom,
+} from './element';
 
 export class BaseElement extends PreactBaseElement {}
 
 /** @override */
-BaseElement['Component'] = BentoVideo;
+BaseElement['Component'] = Component;
 
 /** @override */
-BaseElement['loadable'] = true;
+BaseElement['loadable'] = loadable;
 
 /** @override */
-BaseElement['layoutSizeDefined'] = true;
+BaseElement['layoutSizeDefined'] = layoutSizeDefined;
 
 /**
  * Defaults to `{component: 'video'}` from `BentoVideo` component.
@@ -27,39 +33,10 @@ BaseElement['layoutSizeDefined'] = true;
 BaseElement['staticProps'];
 
 /** @override */
-BaseElement['props'] = {
-  'album': {attr: 'album'},
-  'alt': {attr: 'alt'},
-  'artist': {attr: 'artist'},
-  'artwork': {attr: 'artwork'},
-  'attribution': {attr: 'attribution'},
-  'autoplay': {attr: 'autoplay', type: 'boolean'},
-  'controls': {attr: 'controls', type: 'boolean'},
-  'controlslist': {attr: 'controlslist'},
-  'crossorigin': {attr: 'crossorigin'},
-  'disableremoteplayback': {attr: 'disableremoteplayback'},
-  'loop': {attr: 'loop', type: 'boolean'},
-  'noaudio': {attr: 'noaudio', type: 'boolean'},
-  'poster': {attr: 'poster'},
-  'sources': {
-    selector: 'source',
-    single: false,
-    clone: true,
-  },
-  'src': {attr: 'src'},
-  'title': {attr: 'title'},
-
-  // TODO(alanorozco): These props have no internal implementation yet.
-  'dock': {attr: 'dock', media: true},
-  'rotate-to-fullscreen': {
-    attr: 'rotate-to-fullscreen',
-    type: 'boolean',
-    media: true,
-  },
-};
+BaseElement['props'] = props;
 
 /** @override */
-BaseElement['shadowCss'] = CSS + CSS_AUTOPLAY;
+BaseElement['shadowCss'] = shadowCss;
 
 /** @override */
-BaseElement['usesShadowDom'] = true;
+BaseElement['usesShadowDom'] = usesShadowDom;
