+++ extensions/amp-vimeo/1.0/amp-vimeo.js
@@ -2,14 +2,22 @@ import {isExperimentOn} from '#experiments';
 
 import {userAssert} from '#utils/log';
 
-import {BaseElement} from './base-element';
+import {AmpVideoBaseElement} from 'extensions/amp-video/1.0/video-base-element';
+
+import {
+  Component,
+  layoutSizeDefined,
+  loadable,
+  props,
+  usesShadowDom,
+} from './element';
 
 import {CSS} from '../../../build/amp-vimeo-1.0.css';
 
 /** @const {string} */
 const TAG = 'amp-vimeo';
 
-class AmpVimeo extends BaseElement {
+class AmpVimeo extends AmpVideoBaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -20,6 +28,21 @@ class AmpVimeo extends BaseElement {
     return super.isLayoutSupported(layout);
   }
 }
+
+/** @override */
+AmpVimeo['Component'] = Component;
+
+/** @override */
+AmpVimeo['props'] = props;
+
+/** @override */
+AmpVimeo['layoutSizeDefined'] = layoutSizeDefined;
+
+/** @override */
+AmpVimeo['usesShadowDom'] = usesShadowDom;
+
+/** @override */
+AmpVimeo['loadable'] = loadable;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpVimeo, CSS);
