+++ extensions/amp-vimeo/1.0/element.js
@@ -1,0 +1,17 @@
+export {BentoVimeo as Component} from './component';
+
+export const props = {
+  'autoplay': {attr: 'autoplay', type: 'boolean'},
+  'controls': {attr: 'controls', type: 'boolean'},
+  'videoid': {attr: 'data-videoid'},
+  'doNotTrack': {attr: 'do-not-track'},
+  // TODO(wg-components): Current behavior defaults to loading="auto".
+  // Refactor to make loading="lazy" as the default.
+  'loading': {attr: 'data-loading'},
+};
+
+export const layoutSizeDefined = true;
+
+export const usesShadowDom = true;
+
+export const loadable = true;
