+++ extensions/amp-youtube/1.0/amp-youtube.js
@@ -2,14 +2,16 @@ import {isExperimentOn} from '#experiments';
 
 import {userAssert} from '#utils/log';
 
-import {BaseElement} from './base-element';
+import {AmpVideoBaseElement} from 'extensions/amp-video/1.0/video-base-element';
+
+import {Component, loadable, props, usesShadowDom} from './element';
 
 import {CSS} from '../../../build/amp-youtube-1.0.css';
 
 /** @const {string} */
 const TAG = 'amp-youtube';
 
-class AmpYoutube extends BaseElement {
+class AmpYoutube extends AmpVideoBaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -20,6 +22,18 @@ class AmpYoutube extends BaseElement {
     return super.isLayoutSupported(layout);
   }
 }
+
+/** @override */
+AmpYoutube['Component'] = Component;
+
+/** @override */
+AmpYoutube['loadable'] = loadable;
+
+/** @override */
+AmpYoutube['props'] = props;
+
+/** @override */
+AmpYoutube['usesShadowDom'] = usesShadowDom;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpYoutube, CSS);
